Add files via upload
This is the final project of SKinStore we 4 people have executed this project in 6 days but i am uploading individually cause at that time i havent uploaded anything cause i was not so good with using git hub. SkinStore is a world wide skin care product selling e commerce website.